requirements.txt
matplotlib==3.3.4
imageio
numpy
pillow 
path
scikit-image
